Refactor serialization module: rename 'serialzable' to 'serializable', update documentation, and remove obsolete files
