+++ docs/source/node_editor_api/serialization/index.rst
@@ -4,4 +4,4 @@ Serialization Module
 .. toctree::
    :maxdepth: 1
 
-   serialization
+   serializable
